[IMP]seperation C and B Report
