+++ cb_report/__init__.py
@@ -1,0 +1,3 @@
+from . import reports
+from . import wizard
+from . import controllers
