+++ cb_report/reports/__init__.py
@@ -1,0 +1,1 @@
+from . import report_cb
