+++ cb_report/security/CBreport.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <data>       
+        <record id ='group_cb_report_visibility' model='res.groups'>
+            <field name='name'>C &amp; B report</field>
+            <field name='category_id' ref='base.module_category_hidden'/>
+        </record>
+    </data>
+</odoo>
