+++ customize_vpcs/__init__.py
@@ -1,4 +1,2 @@
 from .import models
 from . import report
-from . import wizard
-from . import controllers
