+++ customize_vpcs/security/ir.model.access.csv
@@ -1,7 +1,6 @@
 id,name,model_id/id,group_id/id,perm_read,perm_write,perm_create,perm_unlink
 access_job_selection,access_job_selection,model_job_selection,base.group_user,1,1,1,1
 access_barged_out,access_barged_out,model_barged_out,base.group_user,1,1,1,1
-report_customize_vpcs_report_cb_report_id,report_customize_vpcs_report_cb_report_id,model_report_customize_vpcs_report_cb_report,customize_vpcs.group_cb_report_visibility,1,1,1,1
 access_custom_tracking_report,access_custom_tracking_report,model_custom_tracking_report,customize_vpcs.group_tracking_report_visibility,1,1,1,1
 access_project_regulate,access_project_regulate,model_project_regulate,base.group_user,1,1,1,1
 access_shipping_line,access_shipping_line,model_shipping_line,base.group_user,1,1,1,1
